refactor: Use header files
