+++ src/games/games.h
@@ -1,8 +1,6 @@
 #ifndef GAMES_H
 #define GAMES_H
 
-#include "blackjack.cpp"
-
 int blackjack();
 
 #endif  // GAMES_H
